+++ spec/unit/resources/matchers_spec.rb
@@ -69,6 +69,9 @@ describe 'oneview_test::default' do
     expect(chef_run).to_not update_oneview_logical_enclosure_from_group('')
     expect(chef_run).to_not reconfigure_oneview_logical_enclosure('')
     expect(chef_run).to_not set_oneview_logical_enclosure_script('')
+    expect(chef_run).to_not create_oneview_logical_enclosure('')
+    expect(chef_run).to_not create_oneview_logical_enclosure_if_missing('')
+    expect(chef_run).to_not delete_oneview_logical_enclosure('')
 
     # oneview_logical_interconnect
     expect(chef_run).to_not add_oneview_logical_interconnect_interconnect('')