+++ CHANGELOG.md
@@ -20,6 +20,8 @@
   - logical_interconnect_group support for API300
   - Deprecate enclosure_group property 'logical_interconnect_group' (string) in favor of 'logical_interconnect_groups' (array)
     - Also supports SAS LIGs for Synergy in this logical_interconnect_groups property
+  - logical_enclosure  support for API300
+  - Add `create`, `create_if_missing` and `delete` actions to logical_enclosure
 
 ### 1.1.0
   - Add support for client ENV variables